Moving middleware to product editor package, outside component
