+++ packages/js/product-editor/package.json
@@ -41,6 +41,7 @@
 		"@woocommerce/number": "workspace:*",
 		"@woocommerce/settings": "^1.0.0",
 		"@woocommerce/tracks": "workspace:^1.3.0",
+		"@wordpress/api-fetch": "wp-6.0",
 		"@wordpress/block-editor": "^9.8.0",
 		"@wordpress/blocks": "^12.3.0",
 		"@wordpress/components": "wp-6.0",
