+++ packages/js/product-editor/src/utils/index.ts
@@ -22,6 +22,7 @@ import { preventLeavingProductForm } from './prevent-leaving-product-form';
 export * from './create-ordered-children';
 export * from './sort-fills-by-order';
 export * from './init-blocks';
+export * from './product-apifetch-middleware';
 
 export {
 	AUTO_DRAFT_NAME,
