+++ plugins/woocommerce-admin/client/products/hooks/use-product-entity-record.ts
@@ -4,19 +4,7 @@
 import { AUTO_DRAFT_NAME } from '@woocommerce/product-editor';
 import { Product } from '@woocommerce/data';
 import { useDispatch, resolveSelect } from '@wordpress/data';
-import { getQuery } from '@woocommerce/navigation';
 import { useEffect, useState } from '@wordpress/element';
-import apiFetch from '@wordpress/api-fetch';
-
-const isProductEditor = () => {
-	const query: { page?: string; path?: string } = getQuery();
-	return (
-		query?.page === 'wc-admin' &&
-		[ '/add-product', '/product/' ].some( ( path ) =>
-			query?.path?.startsWith( path )
-		)
-	);
-};
 
 export function useProductEntityRecord(
 	productId: string | undefined
@@ -25,25 +13,6 @@ export function useProductEntityRecord(
 	const [ product, setProduct ] = useState< Product | undefined >(
 		undefined
 	);
-
-	useEffect( () => {
-		// This is needed to ensure that we use the correct namespace for the entity data store
-		// without disturbing the rest_namespace outside of the product block editor.
-		apiFetch.use( ( options, next ) => {
-			const versionTwoRegex = new RegExp( '^/wp/v2/product' );
-			if (
-				options.path &&
-				versionTwoRegex.test( options?.path ) &&
-				isProductEditor()
-			) {
-				options.path = options.path.replace(
-					versionTwoRegex,
-					'/wc/v3/products'
-				);
-			}
-			return next( options );
-		} );
-	}, [] );
 
 	useEffect( () => {
 		const getRecordPromise: Promise< Product > = productId
